splitted from input_data file benchmark vars
