+++ simba/args/miniproject_config.py
@@ -1,7 +1,7 @@
 from typing import NamedTuple, List
 from pathlib import Path
 
-from simba.args.input_data import BenchmarkInput
+from simba.args.benchmark_input import BenchmarkInput
 from simba.args.toolchain import Toolchain
